requirements.txt
pyyaml
tqdm
numpy
pillow
scikit-learn
pytest
